picui-image-upload-mcp/requirements.txt
httpx>=0.23.0
python-dotenv>=0.21.0
pydantic>=1.10.0 